✅🚚 Move IDLE tests into their own file
This matches the pattern used by most other current "integration tests".
