+++ test/net/imap/test_imap_idle.rb
@@ -1,0 +1,263 @@
+# frozen_string_literal: true
+
+require_relative "../../lib/helper"
+
+class IMAPIdleTest < Net::IMAP::TestCase
+  # TODO: convert to use Net::IMAP::FakeServer::TestHelper
+  include Net::IMAP::TestCase::SimpleTCPServerHelper
+
+  def test_idle
+    server = create_tcp_server
+    port = server.addr[1]
+    requests = []
+    start_server do
+      sock = server.accept
+      begin
+        sock.print("* OK test server\r\n")
+        requests.push(sock.gets)
+        sock.print("+ idling\r\n")
+        sock.print("* 3 EXISTS\r\n")
+        sock.print("* 2 EXPUNGE\r\n")
+        requests.push(sock.gets)
+        sock.print("RUBY0001 OK IDLE terminated\r\n")
+        sock.gets
+        sock.print("* BYE terminating connection\r\n")
+        sock.print("RUBY0002 OK LOGOUT completed\r\n")
+      ensure
+        sock.close
+        server.close
+      end
+    end
+
+    begin
+      imap = Net::IMAP.new(server_addr, :port => port)
+      responses = []
+      imap.idle do |res|
+        responses.push(res)
+        if res.name == "EXPUNGE"
+          imap.idle_done
+        end
+      end
+      assert_equal(3, responses.length)
+      assert_instance_of(Net::IMAP::ContinuationRequest, responses[0])
+      assert_equal("EXISTS", responses[1].name)
+      assert_equal(3, responses[1].data)
+      assert_equal("EXPUNGE", responses[2].name)
+      assert_equal(2, responses[2].data)
+      assert_equal(2, requests.length)
+      assert_equal("RUBY0001 IDLE\r\n", requests[0])
+      assert_equal("DONE\r\n", requests[1])
+      imap.logout
+    ensure
+      imap.disconnect if imap
+    end
+  end
+
+  def test_exception_during_idle
+    server = create_tcp_server
+    port = server.addr[1]
+    requests = []
+    start_server do
+      sock = server.accept
+      begin
+        sock.print("* OK test server\r\n")
+        requests.push(sock.gets)
+        sock.print("+ idling\r\n")
+        sock.print("* 3 EXISTS\r\n")
+        sock.print("* 2 EXPUNGE\r\n")
+        requests.push(sock.gets)
+        sock.print("RUBY0001 OK IDLE terminated\r\n")
+        sock.gets
+        sock.print("* BYE terminating connection\r\n")
+        sock.print("RUBY0002 OK LOGOUT completed\r\n")
+      ensure
+        sock.close
+        server.close
+      end
+    end
+    begin
+      imap = Net::IMAP.new(server_addr, :port => port)
+      begin
+        th = Thread.current
+        m = Monitor.new
+        in_idle = false
+        exception_raised = false
+        c = m.new_cond
+        raiser = Thread.start do
+          m.synchronize do
+            until in_idle
+              c.wait(0.1)
+            end
+          end
+          th.raise(Interrupt)
+          m.synchronize do
+            exception_raised = true
+            c.signal
+          end
+        end
+        @threads << raiser
+        imap.idle do |res|
+          m.synchronize do
+            in_idle = true
+            c.signal
+            until exception_raised
+              c.wait(0.1)
+            end
+          end
+        end
+      rescue Interrupt
+      end
+      assert_equal(2, requests.length)
+      assert_equal("RUBY0001 IDLE\r\n", requests[0])
+      assert_equal("DONE\r\n", requests[1])
+      imap.logout
+    ensure
+      imap.disconnect if imap
+      raiser.kill unless in_idle
+    end
+  end
+
+  def test_idle_done_not_during_idle
+    server = create_tcp_server
+    port = server.addr[1]
+    start_server do
+      sock = server.accept
+      begin
+        sock.print("* OK test server\r\n")
+        sleep 0.1
+      ensure
+        sock.close
+        server.close
+      end
+    end
+    begin
+      imap = Net::IMAP.new(server_addr, :port => port)
+      assert_local_raise(Net::IMAP::Error) do
+        imap.idle_done
+      end
+    ensure
+      imap.disconnect if imap
+    end
+  end
+
+  def test_idle_timeout
+    server = create_tcp_server
+    port = server.addr[1]
+    requests = []
+    start_server do
+      sock = server.accept
+      begin
+        sock.print("* OK test server\r\n")
+        requests.push(sock.gets)
+        sock.print("+ idling\r\n")
+        sock.print("* 3 EXISTS\r\n")
+        sock.print("* 2 EXPUNGE\r\n")
+        requests.push(sock.gets)
+        sock.print("RUBY0001 OK IDLE terminated\r\n")
+        sock.gets
+        sock.print("* BYE terminating connection\r\n")
+        sock.print("RUBY0002 OK LOGOUT completed\r\n")
+      ensure
+        sock.close
+        server.close
+      end
+    end
+
+    begin
+      imap = Net::IMAP.new(server_addr, :port => port)
+      responses = []
+      Thread.pass
+      imap.idle(0.2) do |res|
+        responses.push(res)
+      end
+      # There is no guarantee that this thread has received all the responses,
+      # so check the response length.
+      if responses.length > 0
+        assert_instance_of(Net::IMAP::ContinuationRequest, responses[0])
+        if responses.length > 1
+          assert_equal("EXISTS", responses[1].name)
+          assert_equal(3, responses[1].data)
+          if responses.length > 2
+            assert_equal("EXPUNGE", responses[2].name)
+            assert_equal(2, responses[2].data)
+          end
+        end
+      end
+      # Also, there is no guarantee that the server thread has stored
+      # all the requests into the array, so check the length.
+      if requests.length > 0
+        assert_equal("RUBY0001 IDLE\r\n", requests[0])
+        if requests.length > 1
+          assert_equal("DONE\r\n", requests[1])
+        end
+      end
+      imap.logout
+    ensure
+      imap.disconnect if imap
+    end
+  end
+
+  def test_connection_closed_during_idle
+    server = create_tcp_server
+    port = server.addr[1]
+    requests = []
+    sock = nil
+    threads = []
+    started = false
+    threads << Thread.start do
+      started = true
+      begin
+        sock = server.accept
+        sock.print("* OK test server\r\n")
+        requests.push(sock.gets)
+        sock.print("+ idling\r\n")
+      rescue IOError # sock is closed by another thread
+      ensure
+        server.close
+      end
+    end
+    sleep 0.001 until started
+    threads << Thread.start do
+      imap = Net::IMAP.new(server_addr, :port => port)
+      begin
+        m = Monitor.new
+        in_idle = false
+        closed = false
+        c = m.new_cond
+        threads << Thread.start do
+          m.synchronize do
+            until in_idle
+              c.wait(0.1)
+            end
+          end
+          sock.close
+          m.synchronize do
+            closed = true
+            c.signal
+          end
+        end
+        assert_local_raise(EOFError) do
+          imap.idle do |res|
+            m.synchronize do
+              in_idle = true
+              c.signal
+              until closed
+                c.wait(0.1)
+              end
+            end
+          end
+        end
+        assert_equal(1, requests.length)
+        assert_equal("RUBY0001 IDLE\r\n", requests[0])
+      ensure
+        imap.disconnect if imap
+      end
+    end
+    assert_join_threads(threads)
+  ensure
+    if sock && !sock.closed?
+      sock.close
+    end
+  end
+
+end
